Update print numbers function

+++ 0x10-variadic_functions/1-print_numbers.c
@@ -17,7 +17,7 @@ void print_numbers(const char *separator, const unsigned int n, ...)
 
 	for (i = 0; i < n; i++)
 	{
-		printf("%d", va_arg(list, signed int));
+		printf("%ld", va_arg(list, long int));
 
 		if (separator != NULL && i != n - 1)
 			printf("%s ", separator);
